refactor(allocations): Re-use pagination middleware for allocations
This refactor makes use of the pagination middleware that is used
within the moves app as it is now dynamic enough to be shared to
other collection style listings.
