+++ app/allocations/index.js
@@ -3,39 +3,38 @@ const router = require('express').Router()
 
 const { dateFormat } = require('../../common/helpers/date-utils')
 const { setDateRange, setDatePeriod } = require('../../common/middleware')
+const { setPagination } = require('../../common/middleware/collection')
 const { protectRoute } = require('../../common/middleware/permissions')
 
+const { COLLECTION_PATH, MOUNTPATH } = require('./constants')
 const { dashboard, list } = require('./controllers')
 const {
   setAllocationsByDateAndFilter,
   setAllocationsSummary,
   setAllocationTypeNavigation,
-  setPagination,
 } = require('./middleware')
 
 router.param('period', setDatePeriod)
 router.param('date', setDateRange)
 
-router.get('/', protectRoute('allocations:view'), (req, res, next) => {
+router.use('date', protectRoute('allocations:view'))
+
+router.get('/', (req, res) => {
   const today = format(new Date(), dateFormat)
   return res.redirect(`${req.baseUrl}/week/${today}/`)
 })
+
+router.get('/:period(week|day)/:date/', setAllocationsSummary, dashboard)
+
 router.get(
-  '/:period(week|day)/:date/',
-  protectRoute('allocations:view'),
-  setAllocationsSummary,
-  setPagination,
-  dashboard
-)
-router.get(
-  '/:period(week|day)/:date/:view(outgoing)',
-  protectRoute('allocations:view'),
+  COLLECTION_PATH,
   setAllocationsByDateAndFilter,
   setAllocationTypeNavigation,
-  setPagination,
+  setPagination(MOUNTPATH + COLLECTION_PATH),
   list
 )
+
 module.exports = {
   router,
-  mountpath: '/allocations',
+  mountpath: MOUNTPATH,
 }
