+++ app/allocations/middleware.test.js
@@ -48,144 +48,6 @@ describe('#setAllocationsSummary', function() {
   })
 })
 
-describe('#setPagination', function() {
-  const mockDateRange = ['2020-04-13', '2020-04-19']
-  let req, res, nextSpy
-  beforeEach(function() {
-    this.clock = sinon.useFakeTimers(new Date(mockDateRange[0]).getTime())
-    nextSpy = sinon.spy()
-  })
-
-  afterEach(function() {
-    this.clock.restore()
-  })
-
-  context('by week', function() {
-    beforeEach(function() {
-      res = {
-        locals: {},
-      }
-      req = {
-        baseUrl: '/allocations',
-        params: {
-          date: '2020-04-16',
-          period: 'week',
-          view: 'outgoing',
-        },
-      }
-      nextSpy = sinon.spy()
-      middleware.setPagination(req, res, nextSpy)
-    })
-
-    it('creates pagination on locals', function() {
-      expect(res.locals.pagination).to.exist
-    })
-
-    it('creats correctly todayUrl', function() {
-      expect(res.locals.pagination.todayUrl).to.exist
-      expect(res.locals.pagination.todayUrl).to.equal(
-        '/allocations/week/2020-04-13/outgoing'
-      )
-    })
-
-    it('creats correctly prevUrl', function() {
-      expect(res.locals.pagination.todayUrl).to.exist
-      expect(res.locals.pagination.prevUrl).to.equal(
-        '/allocations/week/2020-04-09/outgoing'
-      )
-    })
-
-    it('creats correctly nextUrl', function() {
-      expect(res.locals.pagination.nextUrl).to.exist
-      expect(res.locals.pagination.nextUrl).to.equal(
-        '/allocations/week/2020-04-23/outgoing'
-      )
-    })
-  })
-  context('by day', function() {
-    beforeEach(function() {
-      res = {
-        locals: {},
-      }
-      req = {
-        baseUrl: '/allocations',
-        params: {
-          date: '2020-04-16',
-          period: 'day',
-        },
-      }
-      nextSpy = sinon.spy()
-      middleware.setPagination(req, res, nextSpy)
-    })
-
-    it('creates pagination on locals', function() {
-      expect(res.locals.pagination).to.exist
-    })
-
-    it('creats correctly todayUrl', function() {
-      expect(res.locals.pagination.todayUrl).to.exist
-      expect(res.locals.pagination.todayUrl).to.equal(
-        '/allocations/day/2020-04-13/'
-      )
-    })
-
-    it('creats correctly prevUrl', function() {
-      expect(res.locals.pagination.prevUrl).to.exist
-      expect(res.locals.pagination.prevUrl).to.equal(
-        '/allocations/day/2020-04-15/'
-      )
-    })
-
-    it('creats correctly nextUrl', function() {
-      expect(res.locals.pagination.nextUrl).to.exist
-      expect(res.locals.pagination.nextUrl).to.equal(
-        '/allocations/day/2020-04-17/'
-      )
-    })
-  })
-  context('with filters', function() {
-    beforeEach(function() {
-      res = {
-        locals: {},
-      }
-      req = {
-        baseUrl: '/allocations',
-        query: {
-          filter1: true,
-          filter2: false,
-        },
-        params: {
-          date: '2020-04-16',
-          period: 'week',
-          view: 'outgoing',
-        },
-      }
-      nextSpy = sinon.spy()
-      middleware.setPagination(req, res, nextSpy)
-    })
-
-    it('creats correctly todayUrl', function() {
-      expect(res.locals.pagination.todayUrl).to.exist
-      expect(res.locals.pagination.todayUrl).to.equal(
-        '/allocations/week/2020-04-13/outgoing?filter1=true&filter2=false'
-      )
-    })
-
-    it('creats correctly prevUrl', function() {
-      expect(res.locals.pagination.todayUrl).to.exist
-      expect(res.locals.pagination.prevUrl).to.equal(
-        '/allocations/week/2020-04-09/outgoing?filter1=true&filter2=false'
-      )
-    })
-
-    it('creats correctly nextUrl', function() {
-      expect(res.locals.pagination.nextUrl).to.exist
-      expect(res.locals.pagination.nextUrl).to.equal(
-        '/allocations/week/2020-04-23/outgoing?filter1=true&filter2=false'
-      )
-    })
-  })
-})
 describe('#setAllocationsByDateAndFilter', function() {
   let next
   beforeEach(async function() {
