+++ common/middleware/collection/index.js
@@ -1,0 +1,5 @@
+const setPagination = require('./set-pagination')
+
+module.exports = {
+  setPagination,
+}
