+++ common/templates/layouts/list.njk
@@ -37,21 +37,23 @@
         </h1>
 
         {% block dateNavigation %}
-          {{ appPagination({
-            classes: "app-pagination--inline app-print--hide",
-            previous: {
-              href: pagination.prevUrl,
-              text: t("actions::previous", { context: period })
-            },
-            next: {
-              href: pagination.nextUrl,
-              text: t("actions::next", { context: period })
-            },
-            items: [{
-              href: pagination.todayUrl,
-              text: t("actions::current", { context: period })
-            }]
-          }) }}
+          {% if pagination %}
+            {{ appPagination({
+              classes: "app-pagination--inline app-print--hide",
+              previous: {
+                href: pagination.prevUrl,
+                text: t("actions::previous", { context: period })
+              },
+              next: {
+                href: pagination.nextUrl,
+                text: t("actions::next", { context: period })
+              },
+              items: [{
+                href: pagination.todayUrl,
+                text: t("actions::current", { context: period })
+              }]
+            }) }}
+          {% endif %}
         {% endblock %}
 
         {% block headerActions %}
